Rename and Reorganise QueryPage
Ticket: #180
