+++ src/app/Main/Body/QueryPage/IntroductionText/QueryPage-IntroductionText.component.tsx
