+++ src/app/Main/Body/QueryPage/QueryPage.component.tsx
@@ -2,7 +2,7 @@
 import { css, jsx } from "@emotion/core";
 import { useState } from "react";
 import { ClippedDrawer as DrawerMenu } from "./Drawer/Drawer-Layout.component";
-import { DrawerControlComponent as DrawerControl} from "./QueryPage-DrawerControl.component";
+import { DrawerControlComponent as DrawerControl} from "./Drawer/ControlBar/QueryPage-DrawerControl.component";
 import { QueryPageContentComponent as QueryPageContent} from "./QueryPage-Content.component";
 
 const mainStyle = css`
